refactor demo a bit
