+++ agents/demo/agent/teacher.py
@@ -1,5 +1,3 @@
-import math
-
 from composabl import Teacher
